Modified association class (ApiParameterLink)
