+++ app/models/__init__.py
@@ -5,5 +5,4 @@ module (as opposed to just their python files)
 
 from .user import *  # noqa
 from .miscellaneous import *  # noqa
-from .parameter import *  # noqa
 from .api import *  # noqa
